test: refactor sampling test
